パソコン上でexeが消されたんで直しましたeset君さぁ…名前だけで消すのはどうかと思うよ？

+++ MapEditor/camera.h
@@ -9,8 +9,8 @@
 
 #include "main.h"
 constexpr double CAMERASPEED(1.0f);
-constexpr double CAMERA_POSV_Y(-400.0f);
-constexpr double CAMERA_POSR_Y(400.0f);
+constexpr float CAMERA_POSV_Y(-400.0f);
+constexpr float CAMERA_POSR_Y(400.0f);
 struct MouseInput {
 	float deltaX, deltaY;
 	float wheelDelta;
